Allow sandboxed JS evaluation over a network connection

+++ src/screen.js
@@ -1,6 +1,6 @@
 const blessed = require('blessed')
 
-exports.createScreen = (client) => {
+exports.createScreen = (client, inputHandler) => {
     var screen = blessed.screen({
         smartCSR: true,
         input: client || process.stdin,
@@ -9,22 +9,71 @@ exports.createScreen = (client) => {
         fullUnicode: true
     })
 
-    screen.key(['C-c', 'q'], (ch, key) => {
+    screen.key([
+        'C-c', 'q'
+    ], (ch, key) => {
         screen.destroy()
     })
 
+    screen.key('i', (key) => {
+        screen
+            .data
+            .main
+            .add('Text')
+    })
+
     screen.on('destroy', () => {
-        if (client.writable) {
+        if (client && client.writable) {
             client.destroy()
         }
     })
 
-    screen.data.main = blessed.box({
-        parent: screen,
-        width: '80%',
-        height: '90%',
-        border: 'line',
-        content: 'Welcome to my server. Here is your own private session.'
+    var body = blessed.box({
+        top: 0,
+        left: 0,
+        height: '100%-1',
+        width: '100%',
+        keys: true,
+        mouse: true,
+        alwaysScroll: true,
+        scrollable: true,
+        scrollbar: {
+            ch: ' ',
+            bg: 'red'
+        }
+    })
+
+    var inputBar = blessed.textbox({
+        bottom: 0,
+        left: 0,
+        height: 1,
+        width: '100%',
+        keys: true,
+        mouse: true,
+        inputOnFocus: true,
+        style: {
+            fg: 'white',
+            bg: 'blue' // Blue background so you see this is different from body
+        }
+    })
+
+    // Add body to blessed screen
+    screen.append(body)
+    screen.append(inputBar)
+
+    const appendText = (text) => {
+        body.pushLine(text)
+        screen.render()
+    }
+
+    inputBar.on('submit', (text) => {
+        appendText(`> ${text}`)
+        appendText(inputHandler(text))
+        inputBar.clearValue()
+    })
+
+    screen.key('enter', (ch, key) => {
+        inputBar.focus();
     })
 
     screen.render()
